+++ src/types/database.ts
@@ -74,6 +74,7 @@ export type Database = {
       }
       church_settings: {
         Row: {
+          address: string | null
           created_at: string
           email_from_name: string | null
           id: string
@@ -86,6 +87,7 @@ export type Database = {
           updated_at: string
         }
         Insert: {
+          address?: string | null
           created_at?: string
           email_from_name?: string | null
           id?: string
@@ -98,6 +100,7 @@ export type Database = {
           updated_at?: string
         }
         Update: {
+          address?: string | null
           created_at?: string
           email_from_name?: string | null
           id?: string